--- test/unit/selectors.test.js
@@ -1,42 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-'use strict'
-
-const { test } = require('tap')
-const { roundRobinSelector, randomSelector } = require('../../lib/Transport').internals
-
-test('RoundRobinSelector', t => {
-  const selector = roundRobinSelector()
-  const arr = [0, 1, 2, 3, 4, 5]
-
-  t.plan(arr.length + 1)
-  for (let i = 0; i <= arr.length; i++) {
-    t.strictEqual(
-      selector(arr),
-      i === arr.length ? arr[0] : arr[i]
-    )
-  }
-})
-
-test('RandomSelector', t => {
-  t.plan(1)
-  const arr = [0, 1, 2, 3, 4, 5]
-  t.type(randomSelector(arr), 'number')
-})
